requirements.txt
streamlit
chromadb
pyyaml
markdown-it-py
llama-index
llama-index-embeddings-openai
llama-index-llms-openai
sentence-transformers
torch
python-dotenv
openai
